feat: move to hooks animation
