+++ app/components/home/portfolio/index.tsx
@@ -1,8 +1,8 @@
 import { Box, Container, Grid, GridItem, Heading, Image } from '@chakra-ui/react'
 import { motion } from 'framer-motion'
 
-import { useAnimations } from '~/constant/useAnimation'
-import { useStaggerAnimation } from '~/constant/useStaggerAnimation'
+import { useAnimations } from '~/hooks/useAnimation'
+import { useStaggerAnimation } from '~/hooks/useStaggerAnimation'
 
 const PortfolioList = [
   {
