+++ app/hooks/useAnimation.ts
