+++ app/hooks/useStaggerAnimation.ts
